Updated storage entities endpoints

+++ src/main/java/com/example/watchmaking/service/StorageService.java
@@ -6,7 +6,6 @@ import com.example.watchmaking.dto.watch.WatchUpdateDto;
 import com.example.watchmaking.entity.Storage;
 import com.example.watchmaking.entity.Watch;
 import com.example.watchmaking.repository.StorageRepository;
-import com.example.watchmaking.repository.WatchRepository;
 import com.example.watchmaking.util.expcetions.InvalidFileTypeException;
 import com.example.watchmaking.util.expcetions.NotFoundException;
 import com.example.watchmaking.util.expcetions.SaveFileException;
@@ -20,6 +19,7 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.*;
+import java.util.stream.Collectors;
 
 @Service
 public class StorageService {
@@ -44,37 +44,80 @@ public class StorageService {
     private static final String IMAGE_BASE_URL = "/images/watch/";
 
     @Transactional
-    public Storage saveFile(UUID watchUuid, StorageInsertDto dto) {
-        MultipartFile file = dto.getFile();
-        try {
-            this.validateFile(file);
-            String originalFileName = StringUtils.cleanPath(file.getOriginalFilename());
-            String uniqueFileName = System.currentTimeMillis() + "_" + originalFileName;
+    public void saveOrUpdateWatchImages(UUID watchUuid, StorageInsertDto dto) {
+        Watch watch = watchService.findByUuid(watchUuid);
+        List<Storage> existingImages = watch.getImages();
 
-            String fileUrl = IMAGE_BASE_URL + uniqueFileName;
+        Set<String> existingOriginalNames = existingImages.stream()
+                .map(storage -> removeTimestampPrefix(storage.getFileName()))
+                .collect(Collectors.toSet());
 
-            storageManager.createDirectoryIfNotExists();
+        Set<String> incomingOriginalNames = dto.getFile().stream()
+                .map(file -> StringUtils.cleanPath(Objects.requireNonNull(file.getOriginalFilename())))
+                .collect(Collectors.toSet());
 
-            Path filePath = storageManager.getUploadPath().resolve(uniqueFileName);
-            Files.copy(file.getInputStream(), filePath);
+        List<Storage> newStorages = new ArrayList<>();
 
-            Storage storage = new Storage(
-                    null,
-                    uniqueFileName,
-                    fileUrl,
-                    file.getContentType(),
-                    file.getSize(),
-                    null
-            );
-            Storage newStorage = storageRepository.save(storage);
-            WatchUpdateDto watchDto = new WatchUpdateDto(watchService.findByUuid(watchUuid));
-            watchDto.setImageUuid(newStorage.getUuid());
-            Watch watch = watchService.updateWatch(watchUuid, newStorage.getUuid(), watchDto);
-            return newStorage;
-        } catch (IOException e) {
-            throw new SaveFileException("Falha ao salvar o arquivo: " + e.getMessage(), e);
+        for (MultipartFile file : dto.getFile()) {
+            if (!file.isEmpty() && !file.getName().isEmpty()) {
+                this.validateFile(file);
+
+
+                String originalFileName = StringUtils.cleanPath(
+                        Objects.requireNonNull(file.getOriginalFilename()).replaceAll("\\s+", "_")
+                );
+
+                if (!existingOriginalNames.contains(originalFileName)) {
+                    String uniqueFileName = System.currentTimeMillis() + "_" + originalFileName;
+                    String fileUrl = IMAGE_BASE_URL + uniqueFileName;
+
+                    try {
+                        storageManager.createDirectoryIfNotExists();
+                        Path filePath = storageManager.getUploadPath().resolve(uniqueFileName);
+                        Files.copy(file.getInputStream(), filePath);
+
+                        Storage storage = new Storage(
+                                null,
+                                uniqueFileName,
+                                fileUrl,
+                                file.getContentType(),
+                                file.getSize(),
+                                null,
+                                watch
+                        );
+                        newStorages.add(storage);
+                    } catch (IOException e) {
+                        throw new SaveFileException("Falha ao salvar o arquivo: " + e.getMessage(), e);
+                    }
+                }
+            }
         }
+
+        List<Storage> imagesToKeep = existingImages.stream()
+                .filter(storage -> incomingOriginalNames.contains(removeTimestampPrefix(storage.getFileName())))
+                .toList();
+
+        List<Storage> imagesToRemove = existingImages.stream()
+                .filter(storage -> !incomingOriginalNames.contains(removeTimestampPrefix(storage.getFileName())))
+                .toList();
+
+        imagesToRemove.forEach((image) -> {
+            this.deleteFile(image.getUuid());
+        });
+
+        List<Storage> finalImages = new ArrayList<>(imagesToKeep);
+        finalImages.addAll(newStorages);
+
+        storageRepository.saveAll(newStorages);
+
+        WatchUpdateDto watchDto = new WatchUpdateDto(watch);
+
+        watchDto.setImages(finalImages);
+
+        watchService.updateWatch(watchUuid, watchDto);
+
     }
+
 
     public Storage findByUuid(UUID uuid) {
         Optional<Storage> optionalStorage = storageRepository.findByUuid(uuid);
@@ -82,38 +125,6 @@ public class StorageService {
             throw new NotFoundException("Arquivo não encontrado");
         }
         return optionalStorage.get();
-    }
-
-    @Transactional
-    public Storage updateFile(UUID uuid, StorageInsertDto dto) {
-        Storage existing = storageRepository.findByUuid(uuid)
-                .orElseThrow(() -> new NotFoundException("Arquivo não encontrado"));
-
-        MultipartFile file = dto.getFile();
-        this.validateFile(file);
-
-        String originalFileName = StringUtils.cleanPath(file.getOriginalFilename());
-        String uniqueFileName = System.currentTimeMillis() + "_" + originalFileName;
-        String fileUrl = IMAGE_BASE_URL + uniqueFileName;
-
-        storageManager.createDirectoryIfNotExists();
-        Path filePath = storageManager.getUploadPath().resolve(uniqueFileName);
-
-        try {
-            Files.copy(file.getInputStream(), filePath);
-
-            if (existing.getFileName() != null) {
-                Path oldFilePath = storageManager.getUploadPath().resolve(existing.getFileName());
-                Files.deleteIfExists(oldFilePath);
-            }
-
-            Storage newStorage = new Storage(existing.getUuid() ,uniqueFileName, fileUrl, file.getContentType(), file.getSize(), null);
-
-            return storageRepository.save(newStorage);
-
-        } catch (IOException e) {
-            throw new SaveFileException("Falha ao atualizar o arquivo: " + e.getMessage(), e);
-        }
     }
 
     @Transactional
@@ -131,14 +142,14 @@ public class StorageService {
     }
 
     private void validateFile(MultipartFile file) {
-        if (file == null || file.isEmpty()) {
-            throw new InvalidFileTypeException("Arquivo não pode ser nulo");
-        }
-
         String contentType = file.getContentType();
-        if (contentType == null || !ALLOWED_CONTENT_TYPES.contains(contentType)) {
+        if (!ALLOWED_CONTENT_TYPES.contains(contentType)) {
             throw new InvalidFileTypeException("Tipo de arquivo não permitido. Apenas imagens são aceitas (JPEG, PNG, GIF, BMP, WEBP)");
         }
     }
 
+    private String removeTimestampPrefix(String fileName) {
+        int underscoreIndex = fileName.indexOf('_');
+        return (underscoreIndex != -1) ? fileName.substring(underscoreIndex + 1) : fileName;
+    }
 }
